+++ hutool-extra/src/main/java/cn/hutool/extra/qrcode/QrCodeUtil.java
@@ -244,6 +244,10 @@ public class QrCodeUtil {
 
 		final HashMap<DecodeHintType, Object> hints = new HashMap<>();
 		hints.put(DecodeHintType.CHARACTER_SET, CharsetUtil.UTF_8);
+		//优化精度
+		hints.put(DecodeHintType.TRY_HARDER, Boolean.TRUE);
+		//复杂模式，开启PURE_BARCODE模式
+		hints.put(DecodeHintType.PURE_BARCODE, Boolean.TRUE);
 		Result result;
 		try {
 			result = formatReader.decode(binaryBitmap, hints);