+++ hutool-extra/src/test/java/cn/hutool/extra/qrcode/QrCodeUtilTest.java
@@ -33,7 +33,7 @@ public class QrCodeUtilTest {
 	}
 
 	@Test
-//	@Ignore
+	@Ignore
 	public void decodeTest() {
 		String decode = QrCodeUtil.decode(FileUtil.file("e:/test.png"));
 		Console.log(decode);